data/gjcl/lsy/trajairnet_led/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py37h89c1867_0
colorama=0.4.4=pyh9f0ad1d_0
cpuonly=1.0=0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
geographiclib=1.52=pyhd8ed1ab_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
intel-openmp=2021.3.0=h06a4308_3350
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
metar=1.8.0=pyhd3deb0d_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
pandas=1.1.3=py37he6710b0_0
pillow=8.3.1=py37h2c7a002_0
pip=21.0.1=py37h06a4308_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.1=py_0
python_abi=3.7=2_cp37m
pytorch=1.9.0=py3.7_cpu_0
pytz=2020.1=py_0
readline=8.1=h27cfd23_0
scipy=1.7.1=py37h292c36d_2
setuptools=52.0.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.10=hbc83047_0
torchaudio=0.9.0=py37
torchvision=0.10.0=py37_cpu
tqdm=4.62.2=pyhd8ed1ab_0
typing_extensions=3.10.0.0=pyhca03da5_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
